WIP - add settings bottom bar
